+++ src/PanConfiguration.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Pan;
 
-/**
- * @internal
- */
 final class PanConfiguration
 {
     /**
@@ -22,6 +19,7 @@ final class PanConfiguration
     private function __construct(
         private int $maxAnalytics = 50,
         private array $allowedAnalytics = [],
+        private string $routePrefix = 'pan',
     ) {
         //
     }
@@ -59,6 +57,16 @@ final class PanConfiguration
     }
 
     /**
+     * Sets the route prefix to be used.
+     *
+     * @internal
+     */
+    public function setRoutePrefix(string $prefix): void
+    {
+        $this->routePrefix = $prefix;
+    }
+
+    /**
      * Sets the maximum number of analytics to store.
      */
     public static function maxAnalytics(int $number): void
@@ -85,6 +93,16 @@ final class PanConfiguration
     }
 
     /**
+     * Sets the route prefix to be used.
+     *
+     * @internal
+     */
+    public static function routePrefix(string $prefix): void
+    {
+        self::instance()->setRoutePrefix($prefix);
+    }
+
+    /**
      * Resets the configuration to its default values.
      *
      * @internal
@@ -93,12 +111,13 @@ final class PanConfiguration
     {
         self::maxAnalytics(50);
         self::allowedAnalytics([]);
+        self::routePrefix('pan');
     }
 
     /**
      * Converts the Pan configuration to an array.
      *
-     * @return array{max_analytics: int, allowed_analytics: array<int, string>}
+     * @return array{max_analytics: int, allowed_analytics: array<int, string>, route_prefix: string}
      *
      * @internal
      */
@@ -107,6 +126,7 @@ final class PanConfiguration
         return [
             'max_analytics' => $this->maxAnalytics,
             'allowed_analytics' => $this->allowedAnalytics,
+            'route_prefix' => $this->routePrefix,
         ];
     }
 }